% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagramme_raoult.R
\name{diagramme_raoult}
\alias{diagramme_raoult}
\title{Diagramme de representaion de la pression partielle en fonction du facteur molaire}
\usage{
diagramme_raoult(x, P0)
}
\arguments{
\item{x}{fraction molaire}

\item{P0}{Pression vapeur du corps pur}
}
\description{
Diagramme de representaion de la pression partielle en fonction du facteur molaire
}
