% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loi_Raoult.R
\name{Pression_partiel}
\alias{Pression_partiel}
\title{Pression partielle selon la loi de Raoult}
\usage{
Pression_partiel(x, P0)
}
\arguments{
\item{x}{Fraction molaire du contituant}

\item{P0}{Pression de vapeur du constituant pur (en Pa ou atm)}
}
\value{
Pression partielle
}
\description{
Pression partielle selon la loi de Raoult
}
\examples{
Pression_partielle(x=0.2,P0= 5)
}
