% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pression_Total.R
\name{Pression_totale}
\alias{Pression_totale}
\title{Pression totale d'une solution idéale}
\usage{
Pression_totale(x_vecteur, P0_vecteur)
}
\arguments{
\item{x_vecteur}{Vecteur de fractions molaires}

\item{P0_vecteur}{Vecteur des pressons de vapeur des corps purs}
}
\value{
Pressin totale
}
\description{
Pression totale d'une solution idéale
}
\examples{
Pression_total(x_vecteur=c(0.2,0,8), P0_vecteur(2,5))

}
